♻️ Refactor api-spec to include error contents and restructure listItem.
The changes mainly focus on adding 'content' to the error codes (400, 404, and 500) providing more details in the case of errors in JSON format. This enhancement offers more specific error information, improving the error debugging process.

Also, it includes reformat. It's now nested with properties under 'items', to improve readability and code structure. Additionally, reformating the 'students' list in the 'LectureDetailResponse'.

This update enhances the clarity and precision of our lecture API.