✨ Set up InversifyJS for dependency injection
Added InversifyJS settings for dependency injection to enhance modularity, testability and maintainability of the app. The Student Controller, Student Service, and Student Repository classes, along with their types and corresponding IOC configurations, have been created. The StudentController has been mapped to the endpoint '/api/students'. A DTO for student creation response was added, and a due getter was added to the student model