♻️ Refactor code for consistency and inline with style guide
This commit involves a series of changes across multiple script files that aim to align the code with the established style guide, thus promoting code readability and maintainability.

In particular, unnecessary blank lines were removed where they were not needed, and semi-colons were added at the end of statements for consistency. In 'student.model.ts', the getter methods for 'nickname' and 'email