🔥 Remove custom env file loader function
The custom function `loadEnvVars` in `env.util.ts` file has been deleted. This custom function was previously used for loading and parsing the environment variables. Due to the availability of more efficient built-in packages such as 'dotenv', the custom function became redundant and thus, was cleared out from the codebase to enhance the efficiency. The deletion was also reflected in `index.ts`