✨ Add Lecture components and configure bindings
A new set of components for Lectures has been added. These include LectureRepository, LectureService, LectureDTOMapper, and LectureController.
The ContainerConfig file was updated to include bindings for these new components for enabling dependency injection. The RouteConfig file was also updated to get a reference to the LectureController. This update allows us to expand the application's functionality for handling lecturing operations.