🔧 Switched test transformation from ts-jest to babel-jest
This change occurred in the jest configuration file. The transformation property has been altered to use 'babel-jest' for '.ts' or '.tsx' files instead of 'ts-jest', and as a result, the need for typescript pre-processing is removed through the jest preset. This could help to improve performance during testing. Additionally, new devDependencies have been added on 'package-lock.json' file, which are necessary to support this babel configuration. These include babel presets and plugins that will be