♻️ Refactor lecture category to string and add validation
The Lecture category has been changed from an 'enum' of predefined values to a simple string type. The reason for this change is to provide more flexibility and to avoid the need for code changes when categories are updated or more categories are added. Introduced validation in the Lecture model to validate the inserted category against existing categories. Moreover, created a new LectureCreateRequest DTO to properly type the incoming POST request in LectureController.