📝 Add 'size' parameter to API and update 'order' enum list
This commit introduces a 'size' query parameter for controlling page sizes in the OpenAPI Specification. It also modifies the 'order' enum list, replacing 'student_count' with 'enrollments'.
The change allows more flexibility when handling paging and ordering of API results and is in line with the updated requirements of the system.