♻️ Refactor createLecture method to use DTO
Two new import statements were added to LectureDTOMapper.ts and a new method, toLectureCreateRequest, was created to map the incoming HTTP request to a LectureCreateRequest.
The createLecture method in LectureService.ts was updated to take a LectureCreateRequest as a parameter rather than individual properties.

In LectureCreateRequest.ts, the class was refactored from an interface to a class with validation decorators and getter methods for its properties.

In LectureController.ts, creation of a LectureCreateRequest object was moved to use the new toLectureCreateRequest method.

Finally, test cases in LectureService.test.ts were updated to account for these changes.

These changes were made to make the code more readable, maintainable and provide a clear separation of concerns.