📝 Refactor query parameters in OpenAPI specification
In light of improving flexibility in search methods, the query parameters have been restructured in the OpenAPI specification. The 'instructor', 'title', and 'studentId' parameters were replaced by 'searchType' and 'searchKeyword'. 'searchType' now allows for choosing the subject of the search and 'searchKeyword' is used to specify the search term. This change supports more flexible and intuitive searches.