♻️ Refactor LectureController unit test for clarity
This commit removes the afterEach block which clears all mocks after each test case, as it was identified that the cleanup in afterEach block was unnecessary. The mockRequest body was also replaced with more explicit values for a better understanding of the test context. Argument checking function was added to validate category in the createLecture function to prevent invalid values. This commits improve the readability and robustness of the code by making it both clear and defensive.