♻️ Refactor tests to use TestLectureFactory
Removed the hardcoded test data used in LectureRepository test and replaced it with the ones created by the TestLectureFactory. This proved to be necessary to increase the maintainability of the tests. The TestLectureFactory creates a standardized set of test data, and this refactoring improves the reusability of the code. This also makes the test results consistent and more reliable as they are now using the same set of test data from the factory. On top of that, the test assertions were simplified, stringing them for id which significantly cleans