♻️ Refactor BaseModel to BaseEntity
In Student, Instructor, Lecture domains and TestLectureFactory, BaseModel has been changed to BaseEntity. We've decided to change the naming from Model to Entity, to accurately represent the nature of these classes and to align with our architecture's terminology. The rename refactor is replaced throughout all instances making the codebase uniform and easier to understand.