📝 Make query parameters optional in OpenAPI schema
Changed the OpenAPI specification for several query parameters by adding the "required": false line. This was achieved to make the parameters optional instead of mandatory in the requests. Additionally, fixed a spelling mistake in "studentId", it is now represented in snake case as "student_id" for consistency with other parameters.