🗃️ Add student counts per lecture and update test data
A new table, lecture_student_counts, was added to keep track of the number of students enrolled in each lecture. This table was added to help the system quickly retrieve the count of students per lecture, instead of performing costly COUNT(*) operations on enrollments.

Test data was updated for enrollments.sql and a new SQL file for test data was created for the aforementioned table.

Docker compose file was adjusted to include the new SQL file and schema.sql was updated to define the new table and index. This was necessary for local testing and ensuring the new table works as expected when deployed.