✨ Add LectureStudentCountRepository to track student count in lectures
Added a new repository, LectureStudentCountRepository, to maintain the count of students enrolled in a particular lecture. This was added to BindingTypes.ts and injected into LectureService and LectureService test cases. Also configured it in ContainerConfig.ts. A new create function is added in LectureService to create student count when a new lecture is added. This will help in providing real-time student count in a lecture