📝 Update OAS document 강의 수정 API parameters and methods
This commit changes the HTTP method from 'put' to 'patch' for updating lectures. The requirement constraint for title, introduction and price has also been removed, giving more flexibility for users when updating lecture information. This is in line with our goal to make the API more forgiving and easy to use.