♻️ Refactor StudentService and related methods for handling non-existing students
A significant shift is made on how non-existing student records are treated across various methods in StudentService and related facades. Previously, "findById" in StudentService used to throw an error if the student does not exist. However, this function now just returns a null. Instead, new method "validateStudentExists" introduced for the explicit purpose of throwing exception for non-existent students.

Also, "findByIdOrReturnDeletedStudent" is changed to "findByIdOrReturnUnknown", creating a placeholder Student record with default 'unknown' values instead of flagged 'deleted' status. This