🔥 Remove findNameById method in InstructorRepository
The findNameById method in the InstructorRepository class has been removed. This method was used to retrieve the name of an instructor based on their ID. However, this functionality has been deemed unnecessary as the instructor's name can be accessed directly from the Instructor object. This change simplifies the codebase and improves efficiency.