🗃️ Refactor SQL query in lecture_student_counts test data file
The SQL query in the lecture_student_counts test data file has been updated for performance improvement. Instead of a simple group operation on active_enrollments, the query now does a left join with the lectures table. This provides a more efficient grouping operation by reducing the size of the dataset to be queried.